requirements.txt
redis>=5.0
azure-core>=1.29.0
azure-cosmos>=4.5.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
panda==0.3.1
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
json==2.0.9